requirements.txt
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
httplib2==0.22.0
idna==3.7
mutagen==1.47.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodomex==3.20.0
pydub==0.25.1
pyfiglet==1.0.2
pyparsing==3.1.2
python-dotenv==1.0.1
pytube==15.0.0
redis==5.0.4
requests==2.32.2
rsa==4.9
setuptools==70.0.0
six==1.16.0
spotipy==2.23.0
uritemplate==4.1.1
urllib3==2.2.1
websockets==12.0
yt-dlp==2024.4.9
